+++ src/main/java/edu/illinois/library/cantaloupe/config/Key.java
@@ -100,6 +100,8 @@ public enum Key {
     HEAPCACHE_TARGET_SIZE("HeapCache.target_size"),
     HTTP_ENABLED("http.enabled"),
     HTTP_HOST("http.host"),
+    HTTP_MAX_THREADS("http.max_threads"),
+    HTTP_MIN_THREADS("http.min_threads"),
     HTTP_PORT("http.port"),
     HTTPSOURCE_ALLOW_INSECURE("HttpSource.allow_insecure"),
     HTTPSOURCE_BASIC_AUTH_SECRET("HttpSource.BasicLookupStrategy.auth.basic.secret"),